requirements.txt
pyglet~=1.5.7
matplotlib~=3.2.2
numpy~=1.19.0
opencv-python~=4.2.0.34

Pillow~=8.0.1